add event cart: added, removed, viewed

+++ src/app/pages/cart/cart-view/cart-view.page.ts
@@ -1,24 +1,24 @@
 import { CommonModule } from '@angular/common';
 import { Component, inject, OnInit } from '@angular/core';
 import { Router, RouterModule } from '@angular/router';
-// prettier-ignore
 import { IonButton, IonContent, IonHeader, IonIcon, IonList, IonTitle, IonToolbar } from '@ionic/angular/standalone';
 import { CartItem } from '../../../interfaces/cart-item';
 import { CartService } from '../../../services/cart.service';
+import { AnalyticsService } from '../../../services/analytic.service';
 
 @Component({
   selector: 'app-cart-view',
   templateUrl: './cart-view.page.html',
   styleUrls: ['./cart-view.page.scss'],
   standalone: true,
-  // prettier-ignore
-  imports: [ IonIcon, IonButton, IonList, IonContent, IonTitle, IonToolbar, IonHeader, CommonModule, RouterModule,
+  imports: [IonIcon, IonButton, IonList, IonContent, IonTitle, IonToolbar, IonHeader, CommonModule, RouterModule,
   ],
 })
 export class CartViewPage implements OnInit {
   // Inyección del servicio
   private cartService = inject(CartService);
   private router = inject(Router);
+  private analyticsService = inject(AnalyticsService);
   // Variables
   cartItems: CartItem[] = [];
   subtotal = 0;
@@ -27,13 +27,15 @@ export class CartViewPage implements OnInit {
 
   ngOnInit() {
     this.loadCartItems();
+    //Evento 10: Cart Viewed 
+    this.analyticsService.track('Cart Viewed');
   }
 
   // Carga los items del carrito
   loadCartItems() {
     this.cartService.getCartItems().subscribe((items) => {
       this.cartItems = items;
-      console.log(this.cartItems); // testing
+      console.log(this.cartItems);
       this.calculateTotals();
     });
   }
@@ -41,8 +43,8 @@ export class CartViewPage implements OnInit {
   // Calcula el total
   calculateTotals() {
     this.subtotal = this.cartItems.reduce(
-      (sum, item) => sum + item.price + item.quantity,
-      0
+      (sum, item) => sum + item.price * item.quantity,
+      0,
     );
     this.total = this.subtotal + this.deliveryFee;
   }
@@ -64,8 +66,22 @@ export class CartViewPage implements OnInit {
   }
 
   // Quitar item del carrito
-  removeItem(itemId: number) {
+  removeItem(itemId: number): void {
+    const item = this.cartItems.find((cartItem) => cartItem.id === itemId);
+
+    if (!item) {
+      return;
+    }
+
     this.cartService.removeItem(itemId);
+
+    // Evento 9: Product Removed
+    this.analyticsService.track('Product Removed', {
+      food_id: item.id,
+      food_name: item.name,
+      price: item.price,
+      quantity: item.quantity,
+    });
   }
 
   // Probar el carrito
